+++ test/td_dd/user_search_filters_test.exs
@@ -1,13 +1,27 @@
 defmodule TdDd.UserSearchFiltersTest do
   use TdDd.DataCase
 
+  import TdDd.TestOperators
+
   alias TdDd.UserSearchFilters
 
   describe "user_search_filters" do
     alias TdDd.UserSearchFilters.UserSearchFilter
 
-    @valid_attrs %{filters: %{}, name: "some name", user_id: 42, scope: :data_structure}
-    @update_attrs %{filters: %{}, name: "some updated name", user_id: 43, scope: :rule}
+    @valid_attrs %{
+      filters: %{},
+      name: "some name",
+      user_id: 42,
+      scope: :data_structure,
+      is_global: true
+    }
+    @update_attrs %{
+      filters: %{},
+      name: "some updated name",
+      user_id: 43,
+      scope: :rule,
+      is_global: false
+    }
     @invalid_attrs %{filters: nil, name: nil, user_id: nil, scope: nil}
 
     def user_search_filter_fixture(attrs \\ %{}) do
@@ -19,9 +33,69 @@ defmodule TdDd.UserSearchFiltersTest do
       user_search_filter
     end
 
-    test "list_user_search_filters/0 returns all user_search_filters" do
+    test "list_user_search_filters/1 returns all user_search_filters" do
       user_search_filter = user_search_filter_fixture()
       assert UserSearchFilters.list_user_search_filters() == [user_search_filter]
+    end
+
+    test "list_user_search_filters/1 filters by scope" do
+      usf = insert(:user_search_filter, scope: "data_structure")
+      insert(:user_search_filter, scope: "rule")
+      assert [usf] <|> UserSearchFilters.list_user_search_filters(%{"scope" => "data_structure"})
+    end
+
+    test "list_user_search_filters/1 filters by user_id" do
+      usf = insert(:user_search_filter, user_id: 1)
+      insert(:user_search_filter, user_id: 2)
+      insert(:user_search_filter, is_global: true)
+      assert [usf] <|> UserSearchFilters.list_user_search_filters(%{"user_id" => 1})
+    end
+
+    test "list_user_search_filters/1 includes global results for all taxonomies" do
+      usf1 = insert(:user_search_filter, user_id: 1)
+      usf2 = insert(:user_search_filter, is_global: true)
+      insert(:user_search_filter, user_id: 2)
+
+      assert [usf1, usf2]
+             <|> UserSearchFilters.list_user_search_filters(%{
+               "user_id" => 1,
+               "with_globals" => :all
+             })
+    end
+
+    test "list_user_search_filters/1 will not duplicate filters" do
+      usf = insert(:user_search_filter, user_id: 1, is_global: true)
+      insert(:user_search_filter, user_id: 2)
+
+      assert [usf]
+             <|> UserSearchFilters.list_user_search_filters(%{
+               "user_id" => 1,
+               "with_globals" => :all
+             })
+    end
+
+    test "list_user_search_filters/1 filters by taxonomy filters on globals" do
+      usf1 = insert(:user_search_filter, user_id: 1)
+      usf2 = insert(:user_search_filter, user_id: 2, is_global: true)
+
+      usf3 =
+        insert(:user_search_filter,
+          user_id: 2,
+          is_global: true,
+          filters: %{"taxonomy" => [1, 2, 3]}
+        )
+
+      insert(:user_search_filter, user_id: 2, filters: %{"taxonomy" => [1, 2, 3]})
+
+      insert(:user_search_filter, user_id: 2, is_global: true, filters: %{"taxonomy" => [7, 8, 9]})
+
+      insert(:user_search_filter, user_id: 2)
+
+      assert [usf1, usf2, usf3]
+             <|> UserSearchFilters.list_user_search_filters(%{
+               "user_id" => 1,
+               "with_globals" => [3, 4, 5]
+             })
     end
 
     test "get_user_search_filter!/1 returns the user_search_filter with given id" do
@@ -39,6 +113,7 @@ defmodule TdDd.UserSearchFiltersTest do
       assert user_search_filter.name == "some name"
       assert user_search_filter.user_id == 42
       assert user_search_filter.scope == :data_structure
+      assert user_search_filter.is_global == true
     end
 
     test "create_user_search_filter/1 with invalid data returns error changeset" do
@@ -61,6 +136,7 @@ defmodule TdDd.UserSearchFiltersTest do
       assert user_search_filter.name == "some updated name"
       assert user_search_filter.user_id == 43
       assert user_search_filter.scope == :rule
+      assert user_search_filter.is_global == false
     end
 
     test "update_user_search_filter/2 with invalid data returns error changeset" do