+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 - [TD-4558] Support for managing data quality functions using API
+- [TD-5195] Support for global user filters
 
 ## [4.52.0] 2022-10-03